% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractMoments.r
\name{binnedMean}
\alias{binnedMean}
\title{Calculate the mean for a matrix of binned data}
\usage{
binnedMean(counts, lowerBounds = seq(0, 95, 5), upperBounds = seq(5, 100,
  5))
}
\arguments{
\item{counts}{}

\item{lowerBounds}{A vector of the lower cut points of the binnned data}

\item{upperBounds}{A vector of the upper cutpoints of the original binnned data}
}
\value{
A vector of estimated means
}
\description{
Calculate the mean for a matrix of binned data
}
\examples{
upper <- seq(10, 100, 10)
lower <- seq(0, 90, 10)
counts2 <- c(1, 2, 1, 4, 4, 2, 0, 0, 0, 0)
counts3 <- c(1, 2, 1, 4, 4, 0, 0, 2, 0, 0)
detectNonContig(counts2)
detectNonContig(counts3)
binnedMean(rbind(counts2, counts3), lower, upper)
}

