% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractMoments.r
\name{binnedMSE}
\alias{binnedMSE}
\title{Calculate the MSE for binned data}
\usage{
binnedMSE(counts, mu, lowerBounds = seq(0, 95, 5), upperBounds = seq(5, 100,
  5))
}
\arguments{
\item{counts}{A matrix of data frame of observed counts of binned data}

\item{mu}{The true observed mean}

\item{lowerBounds}{A vector of the lower cut points of the binnned data}

\item{upperBounds}{A vector of the upper cutpoints of the original binnned data}
}
\value{
A vector with one observation for each row in \code{counts}
}
\description{
Calculate the MSE for binned data
}
\details{
Assumes that the midpoint of the bin is the mean of that prediction.
}
\examples{
upper <- seq(10, 100, 10)
lower <- seq(0, 90, 10)
counts <- c(0, 0, 4, 4, 4, 2, 0, 0, 0, 0)
counts2 <- c(1, 2, 1, 4, 4, 2, 0, 0, 0, 0)
counts_df <- as.data.frame(rbind(counts, counts2))
binnedMSE(counts_df, mu = 33, lowerBounds = lower, upperBounds = upper)
}

