% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractMoments.r
\name{detectNonContig}
\alias{detectNonContig}
\title{Function to find non-contiguity in response patterns}
\usage{
detectNonContig(row, bins = length(row))
}
\arguments{
\item{row}{A vector of binned counts}

\item{bins}{The number of bins: should be the same as the length of \code{row}}
}
\value{
returns either a 0 if there are non-contiguous counts of 1 if there are.
Will often be used with \code{apply} to operate on a matrix
}
\description{
Function to find non-contiguity in response patterns
}
\details{
Assumes there are at least some coins in some bins.
}
\examples{
counts2 <- c(1, 2, 1, 4, 4, 2, 0, 0, 0, 0)
counts3 <- c(1, 2, 0, 4, 4, 2, 0, 0, 0, 0)
detectNonContig(counts2)
detectNonContig(counts3)
apply(rbind(counts2, counts3), 1, detectNonContig)
}

