% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulateBinnningProcess.R
\name{drawUnif}
\alias{drawUnif}
\title{Applies the function replicateBins}
\usage{
drawUnif(count, lowerBound, upperBound, cutSize, repNum)
}
\description{
Applies the function replicateBins
}

