% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatSimOutput.R
\name{formatSimsOutput}
\alias{formatSimsOutput}
\title{Export simulations of re-binning process from \code{rebin} function}
\usage{
formatSimsOutput(simulates, numObs, binLower, binUpper, cutSize, repNum,
  outputType = "list_of_bins")
}
\arguments{
\item{simulates}{a list object created from \code{rebin}}

\item{numObs}{The number of observations that were simulated from}

\item{binLower}{A vector of the lower cut points of the of the original binned ata}

\item{repNum}{The number of replications carried out}

\item{outputType}{currently either \code{"list_of_bins} or \code{list_of_draws}. \code{"list_of_bins} returns a list object
whose length is the number of simulations in each list there is a matrix of counts of how many objects were places in each bin
for each observation. \code{list_of_draws} has a similar data structure but only includes the matrix of draws.}

\item{binUPper}{A vector of the upper cutponts of the original binnned data}

\item{A}{scalar that is the the length of the new (less coarse) simulated data bin}
}
\description{
Export simulations of re-binning process from \code{rebin} function
}
\details{
The idea of this data structure ist that it is similar to an imputation matrix.
}

