% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulateBinnningProcess.R
\name{recutData}
\alias{recutData}
\title{takes data that is drawn from a continous distribution and recuts it.}
\usage{
recutData(simulations, lowerBound, upperBound, cutSize)
}
\description{
takes data that is drawn from a continous distribution and recuts it.
}

