% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulateBinnningProcess.R
\name{replicateBins}
\alias{replicateBins}
\title{For any bin creates replicates from the uniform distribution of coins places in that bin}
\usage{
replicateBins(count, n, lowerBound, upperBound, cutSize)
}
\description{
For any bin creates replicates from the uniform distribution of coins places in that bin
}

