% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sheppardCorrection.r
\name{sheppardCorrection}
\alias{sheppardCorrection}
\title{Calculate the mean of any binned data and the variance and corrected variance}
\usage{
sheppardCorrection(counts, binLower, binUpper)
}
\arguments{
\item{counts}{A row vector the cells associated with count of objects in each bin,}

\item{binLower}{A vector of the lower bounds of each bin}

\item{binUpper}{A vector of the upper bounds of each bin}
}
\value{
A list with three components \enumerate{
 \item The mean of the matrix of binned vectors \code{mu}
 \item The unadjusted variance \code{sigma2}
 \item The variance with the adjusted shepherd correction \code{sigma2Adj}
 }
The the sheppard corrected variance
}
\description{
Calculate the mean of any binned data and the variance and corrected variance
}
\examples{
upper <- seq(10, 100, 10)
lower <- seq(0, 90, 10)
counts <- c(0, 0, 4, 4, 4, 2, 0, 0, 0, 0)
sheppardCorrection(counts, upper, lower)
}

